Introduce event loop infrastructure
Implements a simple event loop that handles timer events. The event loop
is build using the kqueue facility on *BSD/macOS and at the moment is the only
suppored platform.

A big portion of this patch sets up the necessary infrastructure to
format and lint the code, and run automated builds with various
sanitizers.

PR: https://github.com/whalbawi/legatus/pull/2